[react-events] Ensure updateEventListeners updates in commit phase (#16540)
