+++ packages/react-native-renderer/src/ReactNativeHostConfig.js
@@ -501,7 +501,6 @@ export function mountResponderInstance(
   props: Object,
   state: Object,
   instance: Instance,
-  rootContainerInstance: Container,
 ) {
   throw new Error('Not yet implemented.');
 }
